+++ src/main/java/io/github/palexdev/virtualizedfx/flow/paginated/PaginatedVirtualFlowSkin.java
@@ -20,7 +20,8 @@ package io.github.palexdev.virtualizedfx.flow.paginated;
 
 import io.github.palexdev.virtualizedfx.cell.Cell;
 import io.github.palexdev.virtualizedfx.flow.VirtualFlowSkin;
-import javafx.beans.InvalidationListener;
+import javafx.collections.ListChangeListener;
+import javafx.collections.ObservableList;
 import javafx.geometry.Orientation;
 
 /**
@@ -30,24 +31,36 @@ import javafx.geometry.Orientation;
  * {@link PaginatedVirtualFlow#cellSizeProperty()} and of course {@link PaginatedVirtualFlow#orientationProperty()}.
  */
 public class PaginatedVirtualFlowSkin<T, C extends Cell<T>> extends VirtualFlowSkin<T, C> {
-	private InvalidationListener lengthListener;
 
 	//================================================================================
 	// Constructors
 	//================================================================================
 	public PaginatedVirtualFlowSkin(PaginatedVirtualFlow<T, C> virtualFlow) {
 		super(virtualFlow);
-		lengthListener = invalidated -> virtualFlow.updateMaxPage();
-		addListeners();
 	}
 
 	//================================================================================
 	// Methods
 	//================================================================================
 
-	private void addListeners() {
-		PaginatedVirtualFlow<T, C> flow = getFlow();
-		flow.estimatedLengthProperty().addListener(lengthListener);
+	/**
+	 * {@inheritDoc}
+	 * Overridden to also update the number of max pages.
+	 */
+	@Override
+	protected void onListChanged(ObservableList<? extends T> oldList, ObservableList<? extends T> newList) {
+		getFlow().updateMaxPage();
+		super.onListChanged(oldList, newList);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 * Overridden to also update the number of max pages.
+	 */
+	@Override
+	protected void onItemsChanged(ListChangeListener.Change<? extends T> c) {
+		getFlow().updateMaxPage();
+		super.onItemsChanged(c);
 	}
 
 	/**
@@ -119,12 +132,5 @@ public class PaginatedVirtualFlowSkin<T, C extends Cell<T>> extends VirtualFlowS
 		Orientation o = getSkinnable().getOrientation();
 		if (o == Orientation.VERTICAL) return getLength();
 		return super.computeMaxHeight(width, topInset, rightInset, bottomInset, leftInset);
-	}
-
-	@Override
-	public void dispose() {
-		getFlow().estimatedLengthProperty().removeListener(lengthListener);
-		lengthListener = null;
-		super.dispose();
 	}
 }